+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+### Added
+- Implemented a "Delete Question" feature on the question editing page with extra dialog and with tests
+
 ### Updated
 - Clean up connections page and buttons [#516]